pyproject.toml
[tool.poetry]
name = "gemini-bot"
version = "0.1.0"
description = ""
authors = ["hmzi67 <hamzawaheed057@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-generativeai = "^0.4.1"
python-dotenv = "^1.0.1"
streamlit = "^1.32.2"
uvicorn = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
